경매 알림 불러오기 / 경매 list 조회 mapper 수정

+++ src/main/java/com/app/dto/noti/NotificationDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class NotificationDTO {
-    Long notificationId;
+    Integer notificationId;
     String userId;
     String type;
     String message;
